+++ lib/runners/harness.rb
@@ -54,7 +54,11 @@ module Runners
               result = processor.setup do
                 trace_writer.header "Run #{processor.analyzer_name}"
                 @analyzer = processor.analyzer # initialize analyzer
-                exclude_special_dirs { processor.analyze(changes) }
+                if processor.use_git_metadata_dir?
+                  processor.analyze(changes)
+                else
+                  exclude_special_dirs { processor.analyze(changes) }
+                end
               end
 
               case result